# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Player": {
			"properties": {
				"player_id": {
					"type": "number",
					"format": "double"
				},
				"year_played": {
					"type": "string"
				},
				"first_name": {
					"type": "string"
				},
				"last_name": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"pos": {
					"type": "string"
				},
				"num": {
					"type": "string"
				},
				"year": {
					"type": "string"
				},
				"height": {
					"type": "string"
				},
				"height_inches": {
					"type": "number",
					"format": "double"
				},
				"weight": {
					"type": "number",
					"format": "double"
				},
				"hometown": {
					"type": "string"
				},
				"last_school": {
					"type": "string"
				},
				"team": {
					"$ref": "#/definitions/Team"
				},
				"allPurposeRunning": {
					"$ref": "#/definitions/AllPurposeRunning"
				},
				"fumbleReturns": {
					"$ref": "#/definitions/FumbleReturns"
				},
				"interceptions": {
					"$ref": "#/definitions/Interceptions"
				},
				"kickoffReturn": {
					"$ref": "#/definitions/KickoffReturn"
				},
				"kickoffs": {
					"$ref": "#/definitions/Kickoffs"
				},
				"miscDefense": {
					"$ref": "#/definitions/MiscDefense"
				},
				"passing": {
					"$ref": "#/definitions/Passing"
				},
				"placeKicking": {
					"$ref": "#/definitions/PlaceKicking"
				},
				"puntReturn": {
					"$ref": "#/definitions/PuntReturn"
				},
				"punting": {
					"$ref": "#/definitions/Punting"
				},
				"receiving": {
					"$ref": "#/definitions/Receiving"
				},
				"rushing": {
					"$ref": "#/definitions/Rushing"
				},
				"sacks": {
					"$ref": "#/definitions/Sacks"
				},
				"scoring": {
					"$ref": "#/definitions/Scoring"
				},
				"tackles": {
					"$ref": "#/definitions/Tackles"
				},
				"tacklesForLoss": {
					"$ref": "#/definitions/TacklesForLoss"
				},
				"totalOffense": {
					"$ref": "#/definitions/TotalOffense"
				},
				"yardsFromScrimmage": {
					"$ref": "#/definitions/YardsFromScrimmage"
				}
			},
			"required": [
				"player_id",
				"year_played",
				"first_name",
				"last_name",
				"name",
				"pos",
				"num",
				"year",
				"height",
				"height_inches",
				"weight",
				"hometown",
				"last_school",
				"team",
				"allPurposeRunning",
				"fumbleReturns",
				"interceptions",
				"kickoffReturn",
				"kickoffs",
				"miscDefense",
				"passing",
				"placeKicking",
				"puntReturn",
				"punting",
				"receiving",
				"rushing",
				"sacks",
				"scoring",
				"tackles",
				"tacklesForLoss",
				"totalOffense",
				"yardsFromScrimmage"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Team": {
			"properties": {
				"year": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"players": {
					"items": {
						"$ref": "#/definitions/Player"
					},
					"type": "array"
				},
				"conference": {
					"$ref": "#/definitions/Conference"
				}
			},
			"required": [
				"year",
				"name",
				"players",
				"conference"
			],
			"type": "object",
			"additionalProperties": true
		},
...
		"AllPurposeRunning": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"IntRet": {
					"type": "number",
					"format": "double"
				},
				"KickRet": {
					"type": "number",
					"format": "double"
				},
				"Plays": {
					"type": "number",
					"format": "double"
				},
				"PuntRet": {
					"type": "number",
					"format": "double"
				},
				"Recv": {
					"type": "number",
					"format": "double"
				},
				"Rush": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TotalYards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"YardsPlay": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"IntRet",
				"KickRet",
				"Plays",
				"PuntRet",
				"Recv",
				"Rush",
				"Split",
				"TotalYards",
				"YardsG",
				"YardsPlay",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"FumbleReturns": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "number",
					"format": "double"
				},
				"FumRet": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"FumRet",
				"G",
				"Split",
				"TD",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Interceptions": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Int": {
					"type": "number",
					"format": "double"
				},
				"IntG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"Int",
				"IntG",
				"Split",
				"TD",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"KickoffReturn": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Ret": {
					"type": "number",
					"format": "double"
				},
				"RetG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Ret",
				"RetG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Kickoffs": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Kickoffs": {
					"type": "number",
					"format": "double"
				},
				"Onside": {
					"type": "number",
					"format": "double"
				},
				"OutOfBounds": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"Touchback": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Kickoffs",
				"Onside",
				"OutOfBounds",
				"Split",
				"Touchback",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"MiscDefense": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"FumblesForced": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"KicksPuntsBlocked": {
					"type": "number",
					"format": "double"
				},
				"PassesBrokenUp": {
					"type": "number",
					"format": "double"
				},
				"QBHurries": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"FumblesForced",
				"G",
				"KicksPuntsBlocked",
				"PassesBrokenUp",
				"QBHurries",
				"Split",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Passing": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Att": {
					"type": "number",
					"format": "double"
				},
				"AttG": {
					"type": "number",
					"format": "double"
				},
				"Comp": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Int": {
					"type": "number",
					"format": "double"
				},
				"Pct": {
					"type": "number",
					"format": "double"
				},
				"Rating": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsAtt": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Att",
				"AttG",
				"Comp",
				"G",
				"Int",
				"Pct",
				"Rating",
				"Split",
				"TD",
				"Yards",
				"YardsAtt",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"PlaceKicking": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Att": {
					"type": "number",
					"format": "double"
				},
				"AttG": {
					"type": "number",
					"format": "double"
				},
				"ExtraPoint": {
					"type": "number",
					"format": "double"
				},
				"FieldGoal": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Made": {
					"type": "number",
					"format": "double"
				},
				"MadeG": {
					"type": "number",
					"format": "double"
				},
				"Pct": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Att",
				"AttG",
				"ExtraPoint",
				"FieldGoal",
				"G",
				"Made",
				"MadeG",
				"Pct",
				"Split",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"PuntReturn": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Ret": {
					"type": "number",
					"format": "double"
				},
				"RetG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Ret",
				"RetG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Punting": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Punts": {
					"type": "number",
					"format": "double"
				},
				"PuntsG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Punts",
				"PuntsG",
				"Split",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Receiving": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Rec": {
					"type": "number",
					"format": "double"
				},
				"RecG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Rec",
				"RecG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
...
		"PaginatedPlayer": {
			"properties": {
				"players": {
					"items": {
						"$ref": "#/definitions/Player"
					},
					"type": "array"
				},
				"count": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"players",
				"count"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Representation": {
			"properties": {
				"player_id": {
					"type": "number",
					"format": "double"
				},
				"year_played": {
					"type": "string"
				},
				"allPurposeRunning": {
					"type": "boolean"
				},
				"fumbleReturns": {
					"type": "boolean"
				},
				"interceptions": {
					"type": "boolean"
				},
				"kickoffReturn": {
					"type": "boolean"
				},
				"kickoffs": {
					"type": "boolean"
				},
				"miscDefense": {
					"type": "boolean"
				},
				"passing": {
					"type": "boolean"
				},
				"placeKicking": {
					"type": "boolean"
				},
				"puntReturn": {
					"type": "boolean"
				},
				"punting": {
					"type": "boolean"
				},
				"receiving": {
					"type": "boolean"
				},
				"rushing": {
					"type": "boolean"
				},
				"sacks": {
					"type": "boolean"
				},
				"scoring": {
					"type": "boolean"
				},
				"tackles": {
					"type": "boolean"
				},
				"tacklesForLoss": {
					"type": "boolean"
				},
				"totalOffense": {
					"type": "boolean"
				},
				"yardsFromScrimmage": {
					"type": "boolean"
				}
			},
			"required": [
				"player_id",
				"year_played",
				"allPurposeRunning",
				"fumbleReturns",
				"interceptions",
				"kickoffReturn",
				"kickoffs",
				"miscDefense",
				"passing",
				"placeKicking",
				"puntReturn",
				"punting",
				"receiving",
				"rushing",
				"sacks",
				"scoring",
				"tackles",
				"tacklesForLoss",
				"totalOffense",
				"yardsFromScrimmage"
			],
			"type": "object",
			"additionalProperties": true
		},
...
	},
	"info": {
		"title": "new-typeorm-project",
		"version": "0.0.1",
		"description": "Awesome project developed with TypeORM."
	},
	"paths": {
		"/players": {
			"get": {
				"operationId": "GetPlayers",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginatedPlayer"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "team_filter",
						"required": false,
						"collectionFormat": "multi",
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "player_filter",
						"required": false,
						"collectionFormat": "multi",
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/players/player/{player_id}": {
			"get": {
				"operationId": "GetPlayerYears",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginatedPlayer"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "player_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/players/conference": {
			"get": {
				"operationId": "GetPlayersByConference",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {},
							"type": "object"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "conference",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "conference_year",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "team_name",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"collectionFormat": "multi",
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/players/team": {
			"get": {
				"operationId": "GetPlayersByTeam",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "retrieved",
						"schema": {
							"$ref": "#/definitions/PaginatedPlayer"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "school",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "year",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/players/stats/{category}": {
			"get": {
				"operationId": "GetListOfPlayerStats",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Player"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/players/stats/{player_id}/{player_year}/{category}": {
			"get": {
				"operationId": "GetCategoryByString",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Player"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "player_id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "path",
						"name": "player_year",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "category",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/players/stats/categories/representation/{player_id}/{player_year}": {
			"get": {
				"operationId": "GetCategories",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Representation"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "player_id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "path",
						"name": "player_year",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/players/{team_year}/{team_name}": {
			"post": {
				"operationId": "UpdatePlayerToIncludePosAndNum",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "team_year",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "team_name",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"name": {
									"type": "string"
								},
								"height": {
									"type": "string"
								},
								"height_inches": {
									"additionalProperties": true,
									"type": "object"
								},
								"hometown": {
									"type": "string"
								},
								"last_school": {
									"type": "string"
								},
								"pos": {
									"type": "string"
								},
								"num": {
									"type": "string"
								}
							},
							"title": "UpdatePlayerToIncludePosAndNumBody",
							"type": "object",
							"required": [
								"name",
								"height",
								"height_inches",
								"hometown",
								"last_school",
								"pos",
								"num"
							]
						}
					}
				]
			},
			"put": {
				"operationId": "AddPlayer",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "inserted"
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "team_name",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "team_year",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"player_id": {
									"type": "number",
									"format": "double"
								},
								"year_played": {
									"type": "number",
									"format": "double"
								},
								"first_name": {
									"type": "string"
								},
								"last_name": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"year": {
									"type": "string"
								},
								"height": {
									"type": "string"
								},
								"height_inches": {
									"additionalProperties": true,
									"type": "object"
								},
								"weight": {
									"additionalProperties": true,
									"type": "object"
								},
								"hometown": {
									"type": "string"
								},
								"last_school": {
									"type": "string"
								}
							},
							"title": "AddPlayerBody",
							"type": "object",
							"required": [
								"player_id",
								"year_played",
								"first_name",
								"last_name",
								"name",
								"year",
								"height",
								"height_inches",
								"weight",
								"hometown",
								"last_school"
							]
						}
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}